Refactor: Extract core functionality into reusable library package
- Create pkg/histree package with reusable library code
- Update main.go to use the new library package
- Add go get installation instructions to README
- Add library usage examples to documentation
- Make project compatible with go get command
